@@ -5,6 +5,7 @@ namespace App\Http\Controllers\API;
 use App\Http\Controllers\Controller;
 use App\repositories\TaskRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 
 class TasksController extends Controller
 {
@@ -63,7 +64,7 @@ class TasksController extends Controller
     public function store(Request $request)
     {
         $formData = $request->all();
-        $validator = \Validator::make($formData, [
+        $validator = Validator::make($formData, [
             'name' => 'required',
             'description' => 'required',
             'project_id' => 'required'
@@ -106,7 +107,7 @@ class TasksController extends Controller
         }
 
         $formData = $request->all();
-        $validator = \Validator::make($formData, [
+        $validator = Validator::make($formData, [
             'name' => 'required',
             'description' => 'required',
             'project_id' => 'required'
