@@ -19,5 +19,10 @@ class TrustProxies extends Middleware
      *
      * @var int
      */
-    protected $headers = Request::HEADER_X_FORWARDED_ALL;
+    protected $headers =
+   Request::HEADER_X_FORWARDED_FOR |
+   Request::HEADER_X_FORWARDED_HOST |
+   Request::HEADER_X_FORWARDED_PORT |
+   Request::HEADER_X_FORWARDED_PROTO |
+   Request::HEADER_X_FORWARDED_AWS_ELB;
 }
