@@ -3,7 +3,7 @@ import * as React from "react";
 const Footer = () => {
   return (
     <div className="mt-5">
-      <p>&copy; 2020 - All rights reserved</p>
+      <p>Ya paseme profe ;n;</p>
     </div>
   );
 };
