@@ -2,25 +2,10 @@ import React from "react";
 function Contact() {
   return (
     <>
-      <h2>Contact Us</h2>
+      <h2>Contact</h2>
       <div>
         <p>
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Dolorem
-          placeat sit omnis laboriosam repellendus corrupti, excepturi ad
-          impedit nemo harum ullam illo, ipsa eveniet dolores maxime
-          consequuntur iure blanditiis aspernatur!
-        </p>
-        <p>
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Dolorem
-          placeat sit omnis laboriosam repellendus corrupti, excepturi ad
-          impedit nemo harum ullam illo, ipsa eveniet dolores maxime
-          consequuntur iure blanditiis aspernatur!
-        </p>
-        <p>
-          Lorem, ipsum dolor sit amet consectetur adipisicing elit. Dolorem
-          placeat sit omnis laboriosam repellendus corrupti, excepturi ad
-          impedit nemo harum ullam illo, ipsa eveniet dolores maxime
-          consequuntur iure blanditiis aspernatur!
+          20150386@aguascalientes.tecnm.mx
         </p>
       </div>
     </>
