@@ -2,17 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-
-/*
-|--------------------------------------------------------------------------
-| API Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register API routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| is assigned the "api" middleware group. Enjoy building your API!
-|
-*/
+use App\Http\Controllers\API\ProjectsController;
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
@@ -21,7 +11,7 @@ Route::middleware('auth:api')->get('/user', function (Request $request) {
 Route::apiResource('projects', 'API\ProjectsController');
 Route::apiResource('tasks', 'API\TasksController');
 
-Route::get('auth/create-token', 'API\Auth\AuthAPIController@createToken');
+Route::get('auth/create-token', [ProjectsController::class,'createToken']);
 
-Route::post('auth/login', 'API\Auth\AuthAPIController@login');
-Route::post('auth/register', 'API\Auth\AuthAPIController@register');
+Route::post('auth/login', [ProjectsController::class,'login']);
+Route::post('auth/register', [ProjectsController::class,'register']);
